requirements.txt
fastapi
uvicorn
transformers
accelerate
torch
bitsandbytes
langchain
langchain-community
langchain-core
langchain-text-splitters
chromadb
sentence-transformers
pyyaml
gitpython
markdown
